artefy_backend/requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
distlib==0.3.9
Django==5.1.5
django-cors-headers==4.6.0
django-cotton==1.5.2
django-import-export==4.3.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
filelock==3.16.1
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.2.7
pillow==11.1.0
platformdirs==4.3.6
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.0.1
sqlparse==0.5.3
tablib==3.8.0
tzdata==2024.2
virtualenv==20.28.1
